@danielpoe 2019-10-11
Hey jochen - thanks looks good.
Would be perfect if we can measure some differences to see if the change has impact..