@jochen42 2019-10-12
hey @danielpoe,

it would take quite a long time to create an test/messurement with an real-world-impact as an result.
Some other guy's already did some experiments. e.g:

* https://github.com/uber-go/automaxprocs/issues/12#issuecomment-405976401
* https://github.com/coreos/prometheus-operator/issues/501#issuecomment-508865139

But since the results are depending on many different factors (total number of host-cores, configured quoata, count of containers on the host) they don't help us to estimate a real-world impact.

But i will try to explain the technical background and i think i will get clear, that the positive impact will grow, with count of cpu's and the containers of on the host-machine.

If we using docker as the container engine with [cpu quota's] (https://docs.docker.com/config/containers/resource_constraints/#cpu) (kubernetes resource limits are resulting in cpu quota's). The docker runtime (on posix-system's) translates them into cgroup-mounts per container. With these cgroups the linux [cfs-scheduler] (https://developer.ibm.com/tutorials/l-completely-fair-scheduler/) will limit the cpu-time for the docker-container-process and it's childs.
If the container requests to much cpu-time, cfs will throttle those requests -> the process has to wait for computing time -> the process stucks.

If i understand the go scheduler correctly, it distributes the requested go-routines accross os-threads. the maximum amount of threads is limited by GOMAXPROCS. Since one thread can utilize exactly one logical cpu-core and go-routines handles io-wait realy smart, it will utilize each created thread nearly 100%. Now the cpu-quota reduces the amount of available cpu-time via cfs-scheduler and the most cpu-time requests by scheduled go-routines will be throttled.

The included library read's the cfs values of the started process, calculates a virtual amount of cpu-cores and floors it.
Something like this: GOMAXPROCS=Math.floor(cpu.cfs_quota_us / cpu.cfs_period_us). Now the scheduled go-routines schould'nt be throttled anymore.
It will help to reduce cfs throtteling on the host machine.

here a small example which shows the increased cpu-throtteling, if GOMAXPROCS is higher than cpu-limit:

<img width="788" alt="throttletime" src="https://user-images.githubusercontent.com/617145/66691880-e2290000-ec99-11e9-9e2c-f38e20bea6ea.png">



*some more informations*

* articles about the go scheduler
  * https://rakyll.org/scheduler/
  * https://medium.com/@riteeksrivastava/a-complete-journey-with-goroutines-8472630c7f5c
* A nice talk about kubernetes resource-management with many details of cgroups and cfs-scheduler:
  https://www.youtube.com/watch?v=8-apJyr2gi0